pathup/backend/requirements.txt
Flask
pytesseract
scikit-learn
nltk
sentence-transformers
Pillow
PyMuPDF
Flask-Cors